ERP-15 Add Spring-WorkingDay.xml, fix WorkingDay

+++ src/main/com/netcracker/project/model/WorkingDayDAO.java
@@ -1,5 +1,6 @@
 package main.com.netcracker.project.model;
 
+import java.math.BigInteger;
 import java.util.Collection;
 import java.util.Date;
 import main.com.netcracker.project.model.entity.WorkingDay;
@@ -8,5 +9,5 @@ public interface WorkingDayDAO {
 
   void addHoursPerDay(WorkingDay workingDay);
 
-  Collection<WorkingDay> findHoursPerPeriod(Integer userId, Date startDate, Date endDate);
+  Collection<WorkingDay> findHoursPerPeriod(BigInteger userId, Date startDate, Date endDate);
 }
